WebRepositoryEnabledRepository gone in Bamboo 2.5
Not sure if there is an analog now, but WebRepositoryEnabledRepository disappeared
without having previously been deprecated, so I'm not sure what we should be using
instead.  Removing the web repository code from the git plugin enables it to work
with Bamboo 2.5, at least.

Kristian thinks that this feature was just old cruft.